+++ tabling/declarative-programming.md
